pyproject.toml
[project]
name = "ai-eval-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langsmith==0.7.7",
    "python-dotenv>=1.2.1",
]

